feat: full separation of noqx package and solvers
